Update JobInterruptMonitorPlugin.java
Since the Plugin is responsible for more than one job i think it makes sense to store the future not as instance variable in the plugin. 
For me it worked to store the Future in the Job's data map. But i'm not sure if it's the best idea to "pollute" the data map...